requirements.txt
pytrec-eval
cohere==4.36
voyageai
tqdm
vertexai
openai
transformers==4.40
tiktoken
scipy==1.10.1
gritlm
torchmetrics
pyserini
gensim
InstructorEmbedding
google-api-python-client>=2.136.0
google-auth>=2.36.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
